Relocate Pedersen hash personalization enum.
